pyproject.toml
[tool.poetry]
name = "network-vs-atoll"
version = "0.1.0"
description = "Define the cells which are configured on the network but absent in the atoll"
authors = ["Ramil Albakov <ralbakov10@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
cx-Oracle = "^8.3.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.16.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
ipython = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
